FILE: moves dtos and mappers to new lib
