+++ libs/menu-matriarch/shared/data-access-dtos/src/index.ts
@@ -1,1 +1,12 @@
-export * from './lib/menu-matriarch-shared-data-access-dtos.module';
+export * from './lib/dtos/dish-dto';
+export * from './lib/dtos/ingredient-dto';
+export * from './lib/dtos/ingredient-type-dto';
+export * from './lib/dtos/meal-dto';
+export * from './lib/dtos/menu-dto';
+export * from './lib/dtos/tag-dto';
+export * from './lib/dtos/user-dto';
+
+export * from './lib/mappers/map-dish-dto-to-dish';
+export * from './lib/mappers/map-ingredient-type-dto-to-ingredient-type';
+export * from './lib/mappers/map-meal-dto-to-meal';
+export * from './lib/mappers/map-menu-dto-to-menu';
