+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/dtos/_schema.interface.ts
