+++ libs/menu-matriarch/shared/data-access-dtos/src/lib/dtos/dish-dto.ts
