--- libs/menu-matriarch/shared/data-access-dtos/src/lib/menu-matriarch-shared-data-access-dtos.module.ts
@@ -1,7 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@NgModule({
-  imports: [CommonModule],
-})
-export class MenuMatriarchSharedDataAccessDtosModule {}
