+++ libs/menu-matriarch/shared/data-access/src/index.ts
@@ -5,19 +5,6 @@ export * from './lib/loading.service';
 export * from './lib/route.enum';
 export * from './lib/router.service';
 
-export * from './lib/dtos/dish-dto';
-export * from './lib/dtos/ingredient-dto';
-export * from './lib/dtos/ingredient-type-dto';
-export * from './lib/dtos/meal-dto';
-export * from './lib/dtos/menu-dto';
-export * from './lib/dtos/tag-dto';
-export * from './lib/dtos/user-dto';
-
-export * from './lib/mappers/map-dish-dto-to-dish';
-export * from './lib/mappers/map-ingredient-type-dto-to-ingredient-type';
-export * from './lib/mappers/map-meal-dto-to-meal';
-export * from './lib/mappers/map-menu-dto-to-menu';
-
 export * from './lib/updaters/batch.service';
 export * from './lib/updaters/dish-update.service';
 export * from './lib/updaters/ingredient-type-update.service';
